Refactor discovery out of random chains
